fix(maui): fixed web app data's incorrect data type of zoom

+++ Roadnik.MAUI/Data/WebAppState.cs
@@ -1,3 +1,3 @@
 namespace Roadnik.Data;
 
-internal record WebAppState(LatLon Location, bool AutoPan, string? MapLayer, int Zoom);
+internal record WebAppState(LatLon Location, bool AutoPan, string? MapLayer, double Zoom);
